requirements.txt
fastapi[standard]
uvicorn
keras
tensorflow
opencv-python
numpy
pillow